Changed double to float

+++ src/Camera.cc
@@ -76,8 +76,8 @@ Camera::SetHeading(const int heading)
 void
 Camera::Move(const int delta)
 {
-  position.Adjust((int)((double)delta * orientation.GetSin()),
-                  (int)((double)delta * orientation.GetCos()));
+  position.Adjust((int)((float)delta * orientation.GetSin()),
+                  (int)((float)delta * orientation.GetCos()));
   Notify();
 }
